refactor: Extract match creation to dedicated screen with enhanced functionality
- Move match creation/editing from modal to dedicated screen (createMatch.tsx)
- Simplify coachMatches.tsx by removing inline modal and form state
- Add match edit functionality with pre-populated form fields
- Preserve match IDs (home_team_id, away_team_id, location_id, start_time) in Redux state
- Enhance DateTimeSelector with minimum time validation and iOS Done button
- Support external team creation within match flow
- Improve UX with better navigation and haptic feedback
